moved timestamp logic to mod timestamp, introduced fn `apply_transaction` for pieces
